requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv

# HTTP client for API calls
aiohttp==3.9.0
httpx==0.25.2

# Redis for caching
redis==5.0.1

# Environment and configuration
python-dotenv==1.0.0

# Data processing
python-json-logger==2.0.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production server
gunicorn==21.2.0

# Static file serving
aiofiles==23.2.1

# CORS middleware
python-multipart==0.0.6

# Monitoring and logging
structlog==23.2.0

# Optional: For more advanced features
# websockets==12.0  # If you want WebSocket support
# prometheus-client==0.19.0  # For metrics
# sentry-sdk[fastapi]==1.38.0  # For error tracking
